refactor(senpi-task): centralize model registry parsing
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
